+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Released versions should use the Git tag as the section title, for example
 
 ## [Unreleased]
 
+### Changed
+
+- Updated dashboard to show "Check statistics" instead of "Verify outcomes" with better organized information.
+
 ## [v0.2.0] - 2026-06-03
 
 ### Added